+++ ropetest/conftest.py
@@ -14,8 +14,22 @@ def project():
 
 
 @pytest.fixture
+def project2():
+    project = testutils.sample_project("another_project")
+    yield project
+    testutils.remove_project(project)
+
+
+@pytest.fixture
 def project_path(project):
     yield pathlib.Path(project.address)
+
+
+@pytest.fixture
+def project2():
+    project = testutils.sample_project("sample_project2")
+    yield project
+    testutils.remove_project(project)
 
 
 """