+++ ropetest/contrib/autoimport/autoimporttest.py
@@ -1,3 +1,5 @@
+import sqlite3
+from concurrent.futures import ThreadPoolExecutor
 from contextlib import closing, contextmanager
 from textwrap import dedent
 from unittest.mock import ANY, patch
@@ -24,6 +26,20 @@ def is_in_memory_database(connection):
 
 def database_list(connection):
     return list(connection.execute("PRAGMA database_list"))
+
+
+def test_in_memory_database_share_cache(project, project2):
+    ai_1 = AutoImport(project, memory=True)
+    ai_2 = AutoImport(project, memory=True)
+
+    ai_3 = AutoImport(project2, memory=True)
+
+    with ai_1.connection:
+        ai_1.connection.execute("CREATE TABLE shared(data)")
+        ai_1.connection.execute("INSERT INTO shared VALUES(28)")
+    assert ai_2.connection.execute("SELECT data FROM shared").fetchone() == (28,)
+    with pytest.raises(sqlite3.OperationalError, match="no such table: shared"):
+        ai_3.connection.execute("SELECT data FROM shared").fetchone()
 
 
 def test_autoimport_connection_parameter_with_in_memory(
@@ -85,6 +101,37 @@ def test_init_py(
     assert [("from pkg1 import foo", "foo")] == results
 
 
+def test_multithreading(
+    autoimport: AutoImport,
+    project: Project,
+    pkg1: Folder,
+    mod1: File,
+):
+    mod1_init = pkg1.get_child("__init__.py")
+    mod1_init.write(dedent("""\
+        def foo():
+            pass
+    """))
+    mod1.write(dedent("""\
+        foo
+    """))
+    autoimport = AutoImport(project, memory=False)
+    autoimport.generate_cache([mod1_init])
+
+    tp = ThreadPoolExecutor(1)
+    results = tp.submit(autoimport.search, "foo", True).result()
+    assert [("from pkg1 import foo", "foo")] == results
+
+
+def test_connection(project: Project, project2: Project):
+    ai1 = AutoImport(project)
+    ai2 = AutoImport(project)
+    ai3 = AutoImport(project2)
+
+    assert ai1.connection is not ai2.connection
+    assert ai1.connection is not ai3.connection
+
+
 @contextmanager
 def assert_database_is_reset(conn):
     conn.execute("ALTER TABLE names ADD COLUMN deprecated_column")
@@ -139,3 +186,39 @@ def test_setup_db_metadata_table_is_current(autoimport):
     with assert_database_is_preserved(conn), \
             patch("rope.base.versioning.calculate_version_hash", return_value="up-to-date-value"):
         autoimport._setup_db()
+
+
+class TestQueryUsesIndexes:
+    def explain(self, autoimport, query):
+        explanation = list(autoimport._execute(query.explain(), ("abc",)))[0][-1]
+        # the explanation text varies, on some sqlite version
+        explanation = explanation.replace("TABLE ", "")
+        return explanation
+
+    def test_search_by_name_uses_index(self, autoimport):
+        query = models.Name.search_by_name.select_star()
+        assert (
+            self.explain(autoimport, query)
+            == "SEARCH names USING INDEX names_name (name=?)"
+        )
+
+    def test_search_by_name_like_uses_index(self, autoimport):
+        query = models.Name.search_by_name_like.select_star()
+        assert (
+            self.explain(autoimport, query)
+            == "SEARCH names USING INDEX names_name_nocase (name>? AND name<?)"
+        )
+
+    def test_search_module_like_uses_index(self, autoimport):
+        query = models.Name.search_module_like.select_star()
+        assert (
+            self.explain(autoimport, query)
+            == "SEARCH names USING INDEX names_module_nocase (module>? AND module<?)"
+        )
+
+    def test_search_submodule_like_uses_index(self, autoimport):
+        query = models.Name.search_submodule_like.select_star()
+        assert (
+            self.explain(autoimport, query)
+            == "SCAN names" # FIXME: avoid full table scan
+        )